+++ ui-sketcher-webview/.eslintrc.cjs
@@ -11,6 +11,11 @@ module.exports = {
   plugins: ["react-refresh"],
   rules: {
     "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/ban-ts-comment": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      { argsIgnorePattern: "^_", varsIgnorePattern: "^_" },
+    ],
     "react-refresh/only-export-components": [
       "warn",
       { allowConstantExport: true },