+++ ui-sketcher-webview/src/global.d.ts
@@ -1,0 +1,7 @@
+interface Window {
+  displayPreviewOnStart?: boolean;
+  partialRenderEnabled?: boolean;
+  previewUrl?: string;
+  relativePath: string;
+  workspaceId: string;
+}